abstracted operations
also added pretty printing by iterating over the whole tree
